Fix insane bugs (see details)
onblur() is called when a Element is re-appended to the DOM. This calls my wrapped onblur() function and creates new DOM update events during an existing DOM update! Additionally, using toRefs() on an object makes the caller dependent on the assignment of "undefined" to properties. Because  mouseSelectionCandidate could be undefined, this was forcing the ruleView / ruleBody to execute an update.